+++ changelog.md
@@ -4,6 +4,8 @@ Codion Change Log
 ## 0.18.76
 ### is.codion.framework.domain
 - AbstractValueAttributeDefinition bug fixed, now sets false as the default value for non-nullable boolean attributes, when no default value is specified.
+### is.codion.framework.model
+- AbstractEntityEditor bug fixed, clear now sets null for all attributes, modified now works as expected.
 
 ## 0.18.75
 ### is.codion.common.reactive